# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  collection.CreateCollectionRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
  collection.DeleteCollectionsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
  collection.UpdateCollectionRequest:
    properties:
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
...
  /collections:
    delete:
      consumes:
      - application/json
      description: Deletes multiple collections by their IDs
      parameters:
      - description: List of IDs
        in: body
        name: ids
        required: true
        schema:
          $ref: '#/definitions/collection.DeleteCollectionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete multiple collections
      tags:
      - collections
    get:
      description: Returns a list of collections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get collection list
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Creates a new collection and returns the created collection
      parameters:
      - description: Create collection data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/collection.CreateCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new collection
      tags:
      - collections
  /collections/{id}:
    get:
      description: Returns a collection by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a collection
      tags:
      - collections
    put:
      consumes:
      - application/json
      description: Updates a collection
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/collection.UpdateCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a collection
      tags:
      - collections
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"